// backend/src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  DONOR: 'DONOR',
  HOSPITAL: 'HOSPITAL',
  ADMIN: 'ADMIN',
  SUPER_ADMIN: 'SUPER_ADMIN'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const BloodType = {
  A_POSITIVE: 'A_POSITIVE',
  A_NEGATIVE: 'A_NEGATIVE',
  B_POSITIVE: 'B_POSITIVE',
  B_NEGATIVE: 'B_NEGATIVE',
  AB_POSITIVE: 'AB_POSITIVE',
  AB_NEGATIVE: 'AB_NEGATIVE',
  O_POSITIVE: 'O_POSITIVE',
  O_NEGATIVE: 'O_NEGATIVE'
} as const

export type BloodType = (typeof BloodType)[keyof typeof BloodType]


export const Gender = {
  MALE: 'MALE',
  FEMALE: 'FEMALE'
} as const

export type Gender = (typeof Gender)[keyof typeof Gender]


export const Urgency = {
  NORMAL: 'NORMAL',
  URGENT: 'URGENT',
  CRITICAL: 'CRITICAL'
} as const

export type Urgency = (typeof Urgency)[keyof typeof Urgency]


export const RequestStatus = {
  PENDING: 'PENDING',
  MATCHED: 'MATCHED',
  APPROVED: 'APPROVED',
  FULFILLED: 'FULFILLED',
  REJECTED: 'REJECTED'
} as const

export type RequestStatus = (typeof RequestStatus)[keyof typeof RequestStatus]


export const NotificationStatus = {
  SENT: 'SENT',
  READ: 'READ'
} as const

export type NotificationStatus = (typeof NotificationStatus)[keyof typeof NotificationStatus]